refactor project directory structure towards python package
